Upgrade to v1.3.7.1: VRF, uint16 optimizations, and address updates
Upgrades AltarOfAscension and Party contracts to use VRF-based true random power generation and uint16 type for hero power, improving gas efficiency and consistency. Updates all contract addresses and configuration files to v1.3.7.1, including deployment scripts, subgraph config, and documentation. Adds deployment and full pipeline scripts for both v1.3.7.0 and v1.3.7.1, and updates README and deployment metadata to reflect new versions and addresses.

+++ contracts/current/nft/Party.sol
@@ -87,7 +87,7 @@ contract Party is ERC721, Ownable, ReentrancyGuard, Pausable, ERC721Holder {
         // Verify heroes and calculate total power
         for (uint i = 0; i < _heroIds.length; i++) {
             require(heroContract.ownerOf(_heroIds[i]) == msg.sender, "Party: You do not own all heroes");
-            (, uint256 power) = heroContract.getHeroProperties(_heroIds[i]);
+            (, uint16 power) = heroContract.getHeroProperties(_heroIds[i]);
             totalPower += power;
         }
